Refactoring (#68)
* moved cutominput and textarea to same folder

* Reusing prop information in textarea and custominput

* custom hoom for input and textarea for showing errors

* deleted unecessary import from textarea

* moved the PromptIOBox to ContentPanelPrompts component because it's not reusable in whole application
