+++ frontend/src/components/ContentPanel/ContentPanelHeader/PromptCategoryBox/index.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
-import { StateHook } from '../utils/types';
-import { CustomInput } from './Input';
+import { StateHook } from '../../../../utils/types';
+import { CustomInput } from '../../../Inputs';
 
 interface PromptCategoryBoxProps {
     category: string;
